% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwp.R
\name{absorb}
\alias{absorb}
\title{abosorb}
\usage{
absorb(beta1, beta2)
}
\arguments{
\item{beta1}{longwave extinction coefficient (m)}

\item{beta2}{shortwave extinction coefficient (m)}
}
\value{

}
\description{
abosorb
}
