% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwp.R
\name{bulk_mix}
\alias{bulk_mix}
\title{bulk_mix}
\usage{
bulk_mix(pwp_int, params)
}
\arguments{
\item{ml_index}{the index of the depth of the surface mixed layer after adjustment}

\item{s}{}
}
\value{

}
\description{
bulk_mix
}
