% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwp.R
\name{pwp}
\alias{pwp}
\title{Main PWP Program}
\usage{
pwp(
  met_input_file,
  profile_input_file,
  dt = 900,
  dz = 10,
  days = 300,
  depth = 1000,
  dt_save = 4,
  lat = 60,
  rb = 0.64,
  rg = 0.25,
  rkz = 0,
  beta1 = 0.6,
  beta2 = 20
)
}
\arguments{
\item{met_input_file}{Input metereological data}

\item{profile_input_file}{Input profile data}

\item{dt}{time-step increment (seconds)}

\item{dz}{depth increment (meters)}

\item{days}{the number of days to run}

\item{depth}{the depth to run}

\item{dt_save}{time-step increment for saving to file (multiples of dt)}

\item{lat}{latitude (degrees)}

\item{rb}{critical bulk richardson number}

\item{rg}{critical gradient richardson number}

\item{rkz}{background diffusion}

\item{beta1}{longwave extinction coefficient (m)}

\item{beta2}{shortwave extinction coefficient (m)}
}
\value{

}
\description{
Main PWP Program
}
